황승민 : RaisingActor제작 시작

+++ WinAPI_HSM_PrincessMaker2/GameEngineCore/GameEngineLevel.cpp
@@ -179,7 +179,7 @@ void GameEngineLevel::ActorsRender(float _DeltaTime)
 	{
 		TextOutStart = float4::Zero;
 
-		for (size_t i = 0; i < DebugTexts.size(); i++)
+		for (int i = 0; i < DebugTexts.size(); i++)
 		{
 			HDC ImageDc = GameEngineWindow::GetDoubleBufferImage()->GetImageDC();
 			TextOutA(ImageDc, TextOutStart.ix(), TextOutStart.iy(), DebugTexts[i].c_str(), DebugTexts[i].size());
